Rename StaticMedia to SimpleMedia
