+++ packages/lib-react-components/src/SubjectCard/components/SimpleMedia/index.js
@@ -1,0 +1,1 @@
+export { default } from './SimpleMediaContainer'
