--- packages/lib-react-components/src/SubjectCard/components/StaticMedia/index.js
@@ -1,1 +1,0 @@
-export { default } from './StaticMediaContainer'
