+++ packages/lib-react-components/src/SubjectCard/stories/simple/SubjectCard.application.stories.jsx
@@ -7,7 +7,7 @@ import {
 } from '../SubjectCardStoryData'
 
 const meta = {
-  title: 'Components / SubjectCard / Static / Application',
+  title: 'Components / SubjectCard / Simple / Application',
   component: SubjectCard
 }
