+++ packages/lib-react-components/src/SubjectCard/stories/simple/SubjectCard.video.stories.jsx
@@ -6,7 +6,7 @@ import {
 } from '../SubjectCardStoryData'
 
 const meta = {
-  title: 'Components / SubjectCard / Static / Video',
+  title: 'Components / SubjectCard / Simple / Video',
   component: SubjectCard
 }
